requirements.txt
fastapi[all]~=0.98.0
pymongo==3.11
python-dotenv~=1.0.0